% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{check_groupings}
\alias{check_groupings}
\title{Check groupings in data.frame}
\usage{
check_groupings(dat, valid_groups = c("intron", "condition"))
}
\arguments{
\item{dat}{a \code{data.frame} (or derived class)}

\item{valid_groups}{a character string of groups to group_by}
}
\value{
\code{dat} grouped by valid_groups. Internally groups using
\link{\code{dplyr::group_by}}
}
\description{
Check if data.frame is grouped. If it isn't, group and return grouped
data.frame
}

