% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_low_frags}
\alias{filter_low_frags}
\title{Filter introns with too few unique fragments.}
\usage{
filter_low_frags(obj, min_frags, filter_name = paste0("f_low_count_",
  min_frags))
}
\arguments{
\item{obj}{IntronRetention object}

\item{min_frags}{the minimum number of fragments required to pass the filter
in every sample}

\item{filter_name}{a character string denoting the column name in the final
table}
}
\value{
an IntronRetention object with \code{flat} containing a new column
named \code{filter_name}
}
\description{
Filter introns with too few unique fragments.
}

