% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_perfect_psi}
\alias{filter_perfect_psi}
\title{Perfect psi filter}
\usage{
filter_perfect_psi(obj, digits = 5, filter_name = "f_perf_psi")
}
\description{
Remove things that have "perfect" psi scores. A perfect score is when all
samples contain exactly 0 or 1 retention after rounding.
}

