% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zero_coverage.R
\name{get_batch_intron_zc}
\alias{get_batch_intron_zc}
\title{Get several zero coverage results and aggregate them}
\usage{
get_batch_intron_zc(fnames, sample_names, condition_names, regex = "^chr")
}
\arguments{
\item{fnames}{a character vector containing the path to files}

\item{sample_names}{a character vector containing names for each sample. Note: this should correspond to names that are in the IntronRetention table}

\item{condition_names}{this should correspond to the coditions used in newIntronRetention}
}
\value{
a \code{data.frame} containing all of the samples
}
\description{
Takes a list of zero coverage results and aggregates them into a data.frame
that can be used by an IntronRetention class as a filter.
}

