% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zero_coverage.R
\name{get_intron_zc}
\alias{get_intron_zc}
\title{Read zero coverage data}
\usage{
get_intron_zc(fname, sample_name, regex = "^chr",
  adjust_method = "bonferroni", rename_target_id = TRUE)
}
\arguments{
\item{fname}{a character vector of length one containing the file name}

\item{regex}{a regular expression to filter from. For intron retention, you
usually want to only get the introns. If you want to get everything back,
use the regex \code{"*"}.}

\item{adjust_method}{the method to use for multiple testing}

\item{rename_target_id}{if TRUE, rename 'target_id' to 'intron_extension'}

\item{exp_name}{a character containing the name of the experiment}
}
\value{
a data.frame with columns:
\describe{
 \item{\code{intron_extension}}{the intron_extension coordinates}
 \item{\code{len}}{the length of the target}
 \item{\code{zc_len}}{the length of the region with zero coverage}
 \item{zc_pval}{the raw p-value under the Poisson model}
 \item{zc_qval}{the corrected p-value using method \code{adjust_method}}
}
}
\description{
Read the zero coverage data otuput into a \code{data.frame}.
}

