% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readWrite.R
\name{getCol}
\alias{getCol}
\title{Get the columns from eXpress output}
\usage{
getCol(aList, col, substring_replace = NULL)
}
\arguments{
\item{aList}{a list of xprs data.frames}

\item{col}{the column you wish to extract}

\item{substring_replace}{replace this substring from the target_id}
}
\value{
a data.frame with the particular column from express output
}
\description{
Given a list of data.frames (all containing the same type of layout),
}

