% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{getExpCondition}
\alias{getExpCondition}
\title{Get experimental condition}
\usage{
getExpCondition(obj, ...)
}
\arguments{
\item{obj}{the object}

\item{...}{additional arguments}
}
\description{
Get a filtered data.frame back when specifying a condition and slot
}

