% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{intron_length}
\alias{intron_length}
\title{Get intron lengths from an identifier}
\usage{
intron_length(intron_names)
}
\arguments{
\item{intron_names}{the names of the introns in format 'chrom:start-stop'}
}
\value{
an integer vector of the lengths
}
\description{
Transform an intron identifier into lengths.
}

