% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{intron_null_dist}
\alias{intron_null_dist}
\title{Generate the null distribution}
\usage{
intron_null_dist(flat_grouped, n_samp = 10000, test_stat = mean)
}
\arguments{
\item{flat_grouped}{}

\item{n_samp}{number of samples}

\item{test_stat}{test statistic to use (function)}
}
\value{
a numeric with means
}
\description{
Generate the null distribution
}

