% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{install.packages & library}
\usage{
pack(package, character.only = T, quietly = F, lib.loc = NULL,
  warn.conflicts = T)
}
\description{
mom_gamma function will calculate mom estimate for the shape parameter "k" and the scale parameter "theta" by using given sample.
}
\details{
@param package is package names.
 @return package & library
 @examples #' pack(package=c('sp','plyr'))
 @export
}
