% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{print.IntronRetention}
\alias{print.IntronRetention}
\title{Print an IntronRetention object}
\usage{
\method{print}{IntronRetention}(ir)
}
\arguments{
\item{ir}{an IntronRetention object}
}
\value{
the unchanged object \code{ir} after printing a summary to the
terminal
}
\description{
Print an IntronRetention object
}

