% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.R
\name{retention_test}
\alias{retention_test}
\title{Retention test}
\usage{
retention_test(obj, filtered_df = NULL, n_samp = 10000)
}
\arguments{
\item{obj}{an IntronRetention object}

\item{filtered_df}{a data.frame with columns intron, condition, and f_all.
If NULL, use the filters already computed and merge all of them.}

\item{n_samp}{number of samples to generate the null distribution}
}
\value{
a \code{data.frame} with columns intron, condition, mean_retention,
var_retention, pvalue, and qvalue
}
\description{
Compute the retention test using resampling.
}

