% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retentionTest}
\alias{retentionTest}
\title{Single condition intron-retention test}
\usage{
retentionTest(obj, ...)
}
\arguments{
\item{obj}{the object}

\item{...}{additional arguments}
}
\description{
For a single condition, test whether or not the intron was retained
}

