% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summaryStats}
\alias{summaryStats}
\title{Get summary statistics for an IntronRetention object}
\usage{
summaryStats(obj, ...)
}
\arguments{
\item{obj}{the object}

\item{...}{additional arguments}
}
\description{
Get summary statistics for an IntronRetention object.
}

