% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExpCondition}
\alias{getExpCondition}
\title{Get experimental condition}
\usage{
getExpCondition(obj, ...)
}
\arguments{
\item{obj}{the object}

\item{...}{additional arguments}
}
\description{
Get a filtered data.frame back when specifying a condition and slot
}

