% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{lowExpressionFilter}
\alias{lowExpressionFilter}
\title{Filter low expression}
\usage{
lowExpressionFilter(obj, ...)
}
\arguments{
\item{obj}{the type}

\item{...}{additional arguments}
}
\description{
Filter out low expression features.
}
\examples{
lowExpressionFilter()
}

