% Generated by roxygen2 (4.0.1): do not edit by hand
\name{newIntronRetention}
\alias{newIntronRetention}
\title{Compute intron retention}
\usage{
newIntronRetention(targetExpression, intronToUnion, groups)
}
\arguments{
\item{targetExpression}{a matrix of nothing but expression values and one
column containing 'target_id's}

\item{intronToUnion}{a table with one column labeled 'intron' and another
labeled 'target_id' that gives all the target_ids which map to the intron}

\item{groups}{a vector with the grouping}
}
\value{
a IntronRetention object
}
\description{
Create an IntronRetention object.
}

