% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{summaryStats}
\alias{summaryStats}
\title{Get summary statistics for an IntronRetention object}
\usage{
summaryStats(obj, ...)
}
\arguments{
\item{obj}{the object}

\item{...}{additional arguments}
}
\description{
Get summary statistics for an IntronRetention object.
}

